adding routing to composer. adds a "hashchange" event to the mootools events and binds it to a routing function (this is all done behind the scenes).
routing is based off of regular expressions, and is setup like so:

new Composer.Router({
	'/users/([0-9]+)/edit': ['users', 'edit'],
	...
});

now if you go to "#!/users/666/edit" in the browser, it will call the
"edit" function of the global object "users" with the argument 666.
anything inside the regex group is saved as a parameter to be passed to
the routed funciton in the order it appears in the regex.

note that routing is now set up into new Composer.Router({...}) is
called, although the "hashchange" event will exist whenever composer is
loaded.