copied a lot more over from backbone. got rid of the idea of multiple collections per model, each model can only have one colleciton. multiple seemed like a good idea at the time, but doing any sort of useful syncing with that model would be difficult at best.
i still need to implement url structure, but that shouldn't be too bad
now that there is a 1 to 1 for collecitons and models.

so pretty much everything is set up now (including validation). the only
thing left is syncing and collection/model url structure. i'm really
thinking about NOT baking in a syncing function. if i do i'll just take
backbone's since it works pretty damn well, but most of the work we're
going to be using this framework for is talking directly to a backend
server which we have to assume is on a different url, or we'll just be
using local storage. in other words, we'll never use the syncing
function in-house, so why bother.

if we decide to give Composer an actual release sometime, which we
hopefully will after baking a few apps with it, then we can take a look
at polishing everything up. no point in releasing before we know what
works and what doesn't =]