adding a very simple key/value search syntax to the Collection.select function. Instead of:
list.select(function(user) {
	if(	user.get('name') == 'andrew' &&
		user.get('age') == 24 &&
		user.get('friends') == 0 )
	{
		return true;
	}
});

you can now do:

list.select({
	name: 'andrew',
	age: 24,
	friends: 0
});

This is a very *simplified* version of MongoDB's selection syntax and is
nowhere near as useful. it is only for very simple key/value matching.
eventually it can be expanded (there is probably a library from mongo we
could just steal since it's open source).

this works by building a function from a string and compiling it via
"new Function(...);" instead of creating a function that loops over the
given attributes and running it for every item searched. In other words,
its pretty fast since it pre-compiles the search. bitchin