fixing yet another controller extension screwup. i was calling new base().elements to get the elements from that controller and merge them into the inheriting one, buit didnt realize that instantiating controllers willy-nilly is a really horrible idea. oops.
now, when Controller.extend is called, in the generated class, it sets
controller.events & controller.elements into that class, and if it needs
to be extended again, it does so by looking for those elements within
the base class.

i think a lot of this spawned from the use case where you could do
something like

var Animal = Composer.Controller.extend({...});
var animal = new Animal();

var Pelican = animal.extend({...});

This is a horrible idea and not really in-line with our extension
syntax. I may do a lot of reworking having to do with exporting
instances instead of objects in composer. This should all be backwards
compatible too, but I'm starting to regret this decision. Luckily, its
easily changable.