adding a new equality function (Composer.eq) that replaces the one taken from underscore. There was an instance where _.eq was returning false positives:
  {length: 330} == {length: 330, plays: 1}

The new Composer.eq seeks to be fairly false-leaning, and ignore a lot
of edge cases. ovarall, the code is a bit simpler and cleaner, but i'm
sure there is an edge case or two we left out that will break everything
for everyone...good thing this is beta. JKJKLOLOLOL so far no problems
with our highly-complex app, so i think it's working out so far.