Fixing huge bug in Collection
before when a model was removed from a collection, the collection was
trying to unbind _model_event (the function that watches event on the
model) via

  model.unbind('all', this._model_event.bind(this));

This is WRONG because adding a .bind(this) creates a new function that
doesnt exist in the model's event table, the the event is never actually
removed. what this means is that

1. collections perpetually listened to ALL events from ALL models ever
added to them, regardless of whether or not the model was still in that
collection
2. models that existed in any collection were never garbarge collected
until the collection itself was garbage collected.

this is fixed now, by using named events (with the collection's cid)
instead of functions for event refs.