Standardizing Collection.clear() to use remove()
Before, Collection.clear() was removing models manually from the
collection (not using Collection.remove()). and was also firing 'remove'
events *on the model*. interesting choice, past self.

this has been changed so that when clear()ing a colleciton, each model
in the colleciton is individually remove()ed, and there's no chance of
firing a remove event on a removed model.