Bugfix in Controller.attach
Previously, if you specified a manual {inject: element} to a
controller's instantiation, that element would HAVE to exist in the DOM
for the controller's el to be injected into it. because of this, setting
up a number of controllers under a parent controller before its main
element exists in the DOM did not work.

this has been changed so that if an element is passed to inject, that
element is used verbatim without trying to look it up in the DOM