First revision of v1.0
Lots of fun changes!!

- Including new class system w/o Moo dependency, using it for all class
  defs
- New eventing system, based on github.com/orthecreedence/event-glue.
  This is a drop-on replacement but with more features (ie bind_once)
- Splitting composer.js into model, collection, controller .js.
- Simplifying Composer export system.
- Removed (almost?) all dependencies on Mootools, and have now included
  an adapter for handling DOM-based operations which can use wither
  jQuery selectors or Mootools' Slick.

Also, now including (passing) tests for:

- the util system (enables export system and provides some compatilbity
  for Moo functions the old composer used quite a bit: Object.each,
  Object.clone, ...)
- the class system
- the event fabric layer
- Composer.Model
- Composer.Collection
- Composer.Controller
- Composer.RelationalModel

More to come, but this is a great start for re-organizing, cleaning, and
testing the framework.

Also including simple build script (scripts/build) which will create the
same configration that existed before:

 composer.js
 composer.relational.js
 composer.filtercollection.js
 composer.keyboard.js

This configuration may be changed to just include everything (JSmin,
people!) or to let you choose your modules or something. For now, it's
the same as the long-long ago.