Breaking change in RelationalModel
perviously, when calling toJSON() on a relationalmodel, it would assume
that the data stored in model._data (the model's internal data structure
populated by model.set()) was an exact match of the data in the
underlying relaional structure.

this was wrong and stupid, because it assumed that the relational data
would only be set top-level, but even then failed to take object merging
into account.

this has been fixed such that when you call toJSON() on a relational
model, it now populates the returned data using the actual relation
models/collections, respecting the 'sub.object' dot notation.

so from now on, serialization of relational models is *correct* but also
*different* from old cases. if the old method is desired, pass {raw:
true} to the toJSON() function's options param.