Some nice updates to this.parent() in the class system
previously, this.parent() worked by tracking the current function chain
within the class instance and also tracking how deep in the parent chain
we are.

this worked great, until you call a function recursively, in which case
it will count that as a increment in the parent level even though it's
not really calling the parent.

instead of tracking a level, we're now just tracking the parent
functions themselves (which we have access to when wrapping an extended
function) and pushing it onto a state array, so we can always say that
the function at the END of this array is the next parent function to
call.

not only does this save us a lot of obnoxious issues when we have
strange semi-recursive call chains that use this.parent(), but it's
actually a lot more performant as well.

definitely a good commit that, even if not using this.parent() at all,
will still provide some performance benefits, and if using
this.parent(), will make things work in more complicated cases.