Collection reset fix
Previously, Collection.reset() called with option {upsert: true} would
only add new items and update existing items. now, it removes items that
are missing from the given data. this is a more "correct" way of doing a
`reset` (we get both 'add' & 'remove' events now).

this does change things a bit, especially in the relational model. the
relational model uses model/collection.reset() when updating nested
data, and specifies the {upsert: true} option for collections, meaning
it will be affected by this. on the plus side, it was probably broken
before. a negative though is that people might be relying on the broken
behavior.

updating tests to accomodate this change (ran into the issue in the
relational tests).